+++ docs/source/contents/utils.rst
@@ -14,3 +14,14 @@ check_retrieval_dataframe_format
 download_mock_dataset
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 .. autofunction:: oml.utils.download_mock_dataset.download_mock_dataset
+
+PCA
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+.. autoclass:: oml.utils.misc_torch.PCA
+    :undoc-members:
+    :show-inheritance:
+
+    .. automethod:: __init__
+    .. automethod:: transform
+    .. automethod:: inverse_transform
+    .. automethod:: calc_principal_axes_number
