+++ examples/cars/configs/train_cars.yaml
@@ -38,7 +38,7 @@ metric_args:
   cmc_top_k: [1]
   map_top_k: [5]
   fmr_vals: [1]
-  explained_variance_ths: [0.5, 0.9, 0.99]
+  pfc_explained_variance_ths: [0.5, 0.9, 0.99]
   return_only_main_category: True
   visualize_only_main_category: True
