+++ examples/cars/configs/val_cars.yaml
@@ -28,7 +28,7 @@ metric_args:
   map_top_k: [5]
   precision_top_k: [5]
   fmr_vals: [1]
-  explained_variance_ths: [0.5, 0.9, 0.99]
+  pfc_explained_variance_ths: [0.5, 0.9, 0.99]
   return_only_main_category: False
   visualize_only_main_category: True
