+++ examples/sop/configs/train_sop.yaml
@@ -42,7 +42,7 @@ metric_args:
   cmc_top_k: [1]
   map_top_k: [5]
   fmr_vals: []  # Since SOP is a big dataset you should be careful with increasing of the memory footprint, which is needed to calculate fmr
-  explained_variance_ths: [0.5, 0.9, 0.99]
+  pfc_explained_variance_ths: [0.5, 0.9, 0.99]
   return_only_main_category: True
   visualize_only_main_category: True
